# Conflicts:
#	.gradle/4.10.1/fileHashes/fileHashes.bin
#	.gradle/4.10.1/fileHashes/fileHashes.lock
#	.gradle/4.10.1/fileHashes/resourceHashesCache.bin
#	.gradle/4.10.1/javaCompile/jarAnalysis.bin
#	.gradle/4.10.1/javaCompile/javaCompile.lock
#	.gradle/4.10.1/javaCompile/taskHistory.bin
#	.gradle/4.10.1/taskHistory/taskHistory.bin
#	.gradle/4.10.1/taskHistory/taskHistory.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.idea/caches/build_file_checksums.ser
#	.idea/caches/gradle_models.ser
#	.idea/workspace.xml
#	app/src/main/java/com/secureSetting/SecureSettingsMain.java
#	app/src/main/res/layout/activity_secure_main.xml
#	local.properties

+++ app/src/main/java/com/screenlocker/secure/socket/receiver/NetworkReceiver.java
@@ -153,12 +153,15 @@ public class NetworkReceiver extends BroadcastReceiver {
             Timber.d("label %s", label);
             boolean SecureMarket = intent.getBooleanExtra("SecureMarket", false);
 
-            if (SecureMarket) {
+            if (SecureMarket && !aPackageName.equals(context.getPackageName())) {
                 new Thread(() -> MyApplication.getAppDatabase(context).getDao().deleteOne(aPackageName + label)).start();
                 return;
             }
 
-            new Thread(() -> MyApplication.getAppDatabase(context).getDao().deleteOne(aPackageName + label)).start();
+            if(!aPackageName.equals(context.getPackageName())){
+                new Thread(() -> MyApplication.getAppDatabase(context).getDao().deleteOne(aPackageName + label)).start();
+            }
+
             if (intent.hasExtra("isLast")) {
                 Timber.d("ISLAST AVAILABLE : ");
             }
